requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
html2text==2020.1.16
